requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
starlette==0.27.0
python-dotenv==1.0.0
google-generativeai==0.3.0
mysql-connector-python==8.2.0
jinja2==3.1.2
pypdf==5.4.0
python-docx==1.1.2
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
